WidthLimitedIO

Documentation for WidthLimitedIO.

WidthLimitedIO.IncrementalANSIEscape

WidthLimitedIO.ansi_esc_status

WidthLimitedIO.IncrementalANSIEscape — Type

IncrementalANSIEscape

An @enum collection encoding previously-received escape codes. Their primary purpose is to decide whether we are done rendering the escape code.

See ansi_esc_status for usage details.

WidthLimitedIO.ansi_esc_status — Method

ansi_esc_status(status::IncrementalANSIEscape, c::Char) → status
ansi_esc_status(c::Char) → status

Return a status code indicating our position within an ANSI terminal escape code.

Examples

julia> status = ansi_esc_status('m')   # ordinary character with no prior `status`
NONE::IncrementalANSIEscape = 0

julia> status = ansi_esc_status(status, '')  # start an escape sequence
ESCAPE1::IncrementalANSIEscape = 1

julia> status = ansi_esc_status(status, '[')   # CSI
ESCAPE::IncrementalANSIEscape = 2

julia> status = ansi_esc_status(status, '3')   # can be multiple digits
PARAMETER::IncrementalANSIEscape = 5

julia> status = ansi_esc_status(status, 'm')   # terminating final character
FINAL::IncrementalANSIEscape = 6

julia> status = ansi_esc_status(status, 'm')   # now 'm' is again an ordinary character
NONE::IncrementalANSIEscape = 0